refactor Timer
