+++ packages/react-components/react-toast/src/components/Timer/useTimerStyles.styles.ts
@@ -1,0 +1,14 @@
+import { makeStyles } from '@griffel/react';
+
+export const useStyles = makeStyles({
+  progress: {
+    animationName: {
+      from: {
+        opacity: 0,
+      },
+      to: {
+        opacity: 0,
+      },
+    },
+  },
+});
